# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zeitgeist",
    "version": "0.1.0"
  },
  "paths": {
    "/api/runs": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Start Run",
        "description": "202 rather than 201: the run is queued, and for a queued one there is\nnothing at its URL yet.",
        "operationId": "start_run_api_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRunBody"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{run_id}/topics/{topic_id}/renders": {
      "post": {
        "tags": [
          "generation"
        ],
        "summary": "Create Renders",
        "description": "202, and the rows come back `generating`.\n\nThe rows are written before this returns rather than when the job\nfinishes: the database is authoritative for whether a render exists,\nso a tile with no row behind it would vanish on the next reload. A\n`generating` row's `caption_slots` and \u2014 for an auto render \u2014 its\nrationale are not filled in yet and must not be read.\n\nA bare list rather than an envelope, per `api/schemas.py`: there are no\naggregates to carry.",
        "operationId": "create_renders_api_runs__run_id__topics__topic_id__renders_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "topic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Topic Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/LLMGeneration"
                  },
                  {
                    "$ref": "#/components/schemas/ManualGeneration"
                  }
                ],
                "discriminator": {
                  "propertyName": "mode",
                  "mapping": {
                    "llm": "#/components/schemas/LLMGeneration",
                    "manual": "#/components/schemas/ManualGeneration"
                  }
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RenderRecord"
                  },
                  "title": "Response Create Renders Api Runs  Run Id  Topics  Topic Id  Renders Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/renders/{render_id}": {
      "get": {
        "tags": [
          "renders"
        ],
        "summary": "Read Render",
        "operationId": "read_render_api_renders__render_id__get",
        "parameters": [
          {
            "name": "render_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Render Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "renders"
        ],
        "summary": "Remove Render",
        "description": "204 and no body. The row goes first, then both files.\n\nA render whose PNG is already missing still deletes cleanly: the row is\nwhat makes it exist, and the end state is the same either way.\n\nThe meme count every screen shows is `COUNT(*)` over `renders` at query\ntime, so nothing else has to be told this happened \u2014 which is the\nreason that count is not a column.",
        "operationId": "remove_render_api_renders__render_id__delete",
        "parameters": [
          {
            "name": "render_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Render Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutoOrigin": {
        "properties": {
          "provenance": {
            "type": "string",
            "const": "auto",
            "title": "Provenance",
            "default": "auto"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rationale"
        ],
        "title": "AutoOrigin",
        "description": "A render the model briefed: it chose the template and said why."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LLMGeneration": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "llm",
            "title": "Mode",
            "default": "llm"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "count": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Count",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "LLMGeneration",
        "description": "Ask the model to write `count` briefs.\n\n`template_id` is a template the panel picked as an override. None \u2014 the\npanel's default, \"Let the LLM choose\" \u2014 offers the model the whole\nlibrary and lets it pick per brief, exactly as the generate stage does.\nThe below-the-cut `generate` link posts None too: a ranking row has no\nroom to ask which template."
      },
...
      "ManualGeneration": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "manual",
            "title": "Mode",
            "default": "manual"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "caption_slots": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Caption Slots"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "template_id",
          "caption_slots"
        ],
        "title": "ManualGeneration",
        "description": "Render captions a person typed.\n\nNo `count`: the captions describe exactly one meme, and a count beside\nthem would be a representable state with no meaning. No model call, so\nnothing to explain either \u2014 the row's `ManualOrigin` has no rationale\nfield at all."
      },
      "ManualOrigin": {
        "properties": {
          "provenance": {
            "type": "string",
            "const": "manual",
            "title": "Provenance",
            "default": "manual"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ManualOrigin",
        "description": "A render written by hand. No model call, so no choice to explain."
      },
...
      "QueuedRun": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "position"
        ],
        "title": "QueuedRun",
        "description": "`position` 0 means executing now; higher means waiting behind that\nmany runs, which is what the \"queues it behind\" notice reads."
      },
...
      "RenderRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "topic_id": {
            "type": "string",
            "title": "Topic Id"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "caption_slots": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Caption Slots"
          },
          "origin": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AutoOrigin"
              },
              {
                "$ref": "#/components/schemas/ManualOrigin"
              }
            ],
            "title": "Origin",
            "discriminator": {
              "propertyName": "provenance",
              "mapping": {
                "auto": "#/components/schemas/AutoOrigin",
                "manual": "#/components/schemas/ManualOrigin"
              }
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "generating",
              "ready",
              "failed"
            ],
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "run_id",
          "topic_id",
          "template_id",
          "caption_slots",
          "origin",
          "status",
          "error",
          "created_at"
        ],
        "title": "RenderRecord",
        "description": "One rendered meme. The database is authoritative for whether it exists;\nthe PNG and its thumbnail live at `output/<run_id>/renders/<id>.png`.\n\nWhile `status` is `\"generating\"`, `caption_slots` and `origin`'s\n`rationale` (on `AutoOrigin`) are not yet written \u2014 they hold\nplaceholder values until the job finishes drawing this render \u2014 and\nmust not be read as its real captions or the model's real reasoning.\n\n`template_id` is None where no template has been chosen: a `generating`\nrow whose request left the choice to the model, and a `failed` row\nwhose brief failed before the model chose. That is a real state rather\nthan a placeholder \u2014 there is no template to name \u2014 so it is `| None`\nrather than an empty string standing in for one."
      },
...
      "RunConfig": {
        "properties": {
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "trend_limit": {
            "type": "integer",
            "title": "Trend Limit"
          },
          "posts_per_trend": {
            "type": "integer",
            "title": "Posts Per Trend"
          },
          "top_count": {
            "type": "integer",
            "title": "Top Count"
          },
          "meme_potential_weight": {
            "type": "number",
            "title": "Meme Potential Weight"
          },
          "phrase_min_authors": {
            "type": "integer",
            "title": "Phrase Min Authors"
          },
          "distil_char_budget": {
            "type": "integer",
            "title": "Distil Char Budget"
          },
          "distil_concurrency": {
            "type": "integer",
            "title": "Distil Concurrency"
          },
          "llm_provider": {
            "type": "string",
            "title": "Llm Provider"
          },
          "llm_model": {
            "type": "string",
            "title": "Llm Model"
          },
          "template_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sources",
          "trend_limit",
          "posts_per_trend",
          "top_count",
          "meme_potential_weight",
          "phrase_min_authors",
          "distil_char_budget",
          "distil_concurrency",
          "llm_provider",
          "llm_model",
          "template_ids"
        ],
        "title": "RunConfig",
        "description": "The settings a run used, frozen at its start.\n\nA copy rather than a reference: run detail's config line and the \"Re-run\nconfig\" action must show what the run actually used, which is not\nrecoverable from a `.env` that has since been edited."
      },
...
      "RunError": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "stage": {
            "$ref": "#/components/schemas/Stage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "message",
          "stage"
        ],
        "title": "RunError",
        "description": "Why a run failed, and where.\n\n`kind` is the exception class name rather than the instance, because the\nRuns screen renders it as a label \u2014 \"RenderError in generate\"."
      },
      "RunPage": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Runs"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "runs",
          "next_cursor"
        ],
        "title": "RunPage",
        "description": "One page of runs. `next_cursor` is None on the last page."
      },
      "RunRecordRow": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "ok",
              "failed",
              "aborted",
              "interrupted"
            ],
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "attempt_started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Attempt Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "config": {
            "$ref": "#/components/schemas/RunConfig"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunError"
              },
              {
                "type": "null"
              }
            ]
          },
          "item_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Count"
          },
          "trends_found": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trends Found"
          },
          "topics_kept": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topics Kept"
          },
          "phrases_found": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phrases Found"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "status",
          "started_at",
          "attempt_started_at",
          "finished_at",
          "config",
          "error",
          "item_count",
          "trends_found",
          "topics_kept",
          "phrases_found"
        ],
        "title": "RunRecordRow",
        "description": "A row of `run_records`, read back.\n\nThe counts are `| None` because they are written when the run finishes;\na run still in flight has not counted anything yet."
      },
      "RunSummary": {
        "properties": {
          "run": {
            "$ref": "#/components/schemas/RunRecordRow"
          },
          "topic_labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topic Labels"
          },
          "render_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Render Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run",
          "topic_labels",
          "render_ids"
        ],
        "title": "RunSummary",
        "description": "One row of the Runs list.\n\n`topic_labels` and `render_ids` are joins the row needs and the run\nrecord does not carry: column two is the titles, column four the\nthumbnails."
      },
...
      "Stage": {
        "type": "string",
        "enum": [
          "ingest",
          "analyse",
          "evaluate",
          "generate"
        ],
        "title": "Stage"
      },
...
      "StartRunBody": {
        "properties": {
          "template_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Ids"
          },
          "overrides": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Overrides",
            "default": {}
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "StartRunBody",
        "description": "What `POST /api/runs` accepts.\n\n`overrides` is `str`-valued because the values come from a form and are\nvalidated by `Settings` when the worker freezes the run's config \u2014 the\none place that knows each field's real type."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}